requirements.txt
Django
Celery
redis
django-celery-beat
django-celery-results
dj-database-url
psycopg[binary]
python-decouple
django-timescaledb
gunicorn
kombu
whitenoise
eventlet
django-ninja
